backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Saimon Zinchenko <s.zinchenko@ktsstudio.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
marshmallow = "^3.20.2"
django-serializer = "^1.2.1"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
